requirements.txt
Flask==3.0.2
gunicorn==20.1.0
Werkzeug==3.0.1
google_cloud_aiplatform==1.43.0
langchain==0.1.9
langchain_community==0.0.24
langchain_core==0.1.28
langchain_google_vertexai==0.1.0
protobuf==4.25.3
openai==1.13.3
transformers==4.38.2